rstp: Change transitions order.
Without this patch some IXIA tests fail, but standard 802.1D-2004 says
that the order in which conditions are checked when exiting a state does
not matter.
Where two or more exit conditions with the same level of precedence
become TRUE simultaneously, the choice as to which exit condition causes
the state transition to take place is arbitrary.