stp, rstp: do not forward bpdu packets in STP/RSTP disabled state.
Commit 4b5f19962adccfb3b4a431c9787c45e2c410ab55 made a distinction between a
port with STP/RSTP protocol enabled and a disabled state and a port which has a
disabled state because STP/RSTP is not active.

Commit bacdb85ad82f981697245eefb40a3b360cfe379b was necessary to forward stp
bpdu packets when STP was disabled.

With this patch, stp bpdu packets are forwarded if STP/RSTP is disabled, but
not when the STP/RSTP protocol is enabled and a port has a disabled state.

Signed-off-by: Daniele Venturino <daniele.venturino@m3s.it>